# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Robot Fleet API
  description: Public API for QA engineers to practice API testing, automation,
    contract testing, performance testing, security testing and modern backend
    quality engineering.
  version: 1.0.0
  contact:
    name: Robot Fleet API
    url: https://github.com
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    def-0:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required:
        - page
        - limit
        - total
        - totalPages
      title: PaginationMeta
...
paths:
...
  /api/v1/users/:
    get:
      summary: List users
      tags:
        - Users
      parameters:
        - schema:
            minimum: 1
            default: 1
            type: integer
          in: query
          name: page
          required: false
        - schema:
            minimum: 1
            maximum: 100
            default: 20
            type: integer
          in: query
          name: limit
          required: false
        - schema:
            anyOf:
              - type: string
                enum:
                  - ACTIVE
              - type: string
                enum:
                  - INACTIVE
              - type: string
                enum:
                  - SUSPENDED
          in: query
          name: status
          required: false
        - schema:
            anyOf:
              - type: string
                enum:
                  - ADMIN
              - type: string
                enum:
                  - MANAGER
              - type: string
                enum:
                  - OPERATOR
              - type: string
                enum:
                  - VIEWER
          in: query
          name: role
          required: false
        - schema:
            type: string
          in: query
          name: search
          required: false
        - schema:
            default: createdAt
            type: string
          in: query
          name: sort
          required: false
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          format: uuid
                          type: string
                        name:
                          type: string
                        email:
                          format: email
                          type: string
                        role:
                          anyOf:
                            - type: string
                              enum:
                                - ADMIN
                            - type: string
                              enum:
                                - MANAGER
                            - type: string
                              enum:
                                - OPERATOR
                            - type: string
                              enum:
                                - VIEWER
                        status:
                          anyOf:
                            - type: string
                              enum:
                                - ACTIVE
                            - type: string
                              enum:
                                - INACTIVE
                            - type: string
                              enum:
                                - SUSPENDED
                        createdAt:
                          format: date-time
                          type: string
                        updatedAt:
                          format: date-time
                          type: string
                      required:
                        - id
                        - name
                        - email
                        - role
                        - status
                        - createdAt
                        - updatedAt
                      title: User
                  pagination:
                    $ref: "#/components/schemas/def-0"
                required:
                  - data
                  - pagination
    post:
      summary: Create a user
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  minLength: 2
                  maxLength: 120
                  type: string
                email:
                  format: email
                  type: string
                password:
                  minLength: 8
                  maxLength: 72
                  type: string
                role:
                  anyOf:
                    - type: string
                      enum:
                        - ADMIN
                    - type: string
                      enum:
                        - MANAGER
                    - type: string
                      enum:
                        - OPERATOR
                    - type: string
                      enum:
                        - VIEWER
              required:
                - name
                - email
                - password
        required: true
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    format: uuid
                    type: string
                  name:
                    type: string
                  email:
                    format: email
                    type: string
                  role:
                    anyOf:
                      - type: string
                        enum:
                          - ADMIN
                      - type: string
                        enum:
                          - MANAGER
                      - type: string
                        enum:
                          - OPERATOR
                      - type: string
                        enum:
                          - VIEWER
                  status:
                    anyOf:
                      - type: string
                        enum:
                          - ACTIVE
                      - type: string
                        enum:
                          - INACTIVE
                      - type: string
                        enum:
                          - SUSPENDED
                  createdAt:
                    format: date-time
                    type: string
                  updatedAt:
                    format: date-time
                    type: string
                required:
                  - id
                  - name
                  - email
                  - role
                  - status
                  - createdAt
                  - updatedAt
  /api/v1/users/{id}:
    get:
      summary: Get a user by id
      tags:
        - Users
      parameters:
        - schema:
            format: uuid
            type: string
          in: path
          name: id
          required: true
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    format: uuid
                    type: string
                  name:
                    type: string
                  email:
                    format: email
                    type: string
                  role:
                    anyOf:
                      - type: string
                        enum:
                          - ADMIN
                      - type: string
                        enum:
                          - MANAGER
                      - type: string
                        enum:
                          - OPERATOR
                      - type: string
                        enum:
                          - VIEWER
                  status:
                    anyOf:
                      - type: string
                        enum:
                          - ACTIVE
                      - type: string
                        enum:
                          - INACTIVE
                      - type: string
                        enum:
                          - SUSPENDED
                  createdAt:
                    format: date-time
                    type: string
                  updatedAt:
                    format: date-time
                    type: string
                required:
                  - id
                  - name
                  - email
                  - role
                  - status
                  - createdAt
                  - updatedAt
    put:
      summary: Replace a user
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  minLength: 2
                  maxLength: 120
                  type: string
                email:
                  format: email
                  type: string
                role:
                  anyOf:
                    - type: string
                      enum:
                        - ADMIN
                    - type: string
                      enum:
                        - MANAGER
                    - type: string
                      enum:
                        - OPERATOR
                    - type: string
                      enum:
                        - VIEWER
                status:
                  anyOf:
                    - type: string
                      enum:
                        - ACTIVE
                    - type: string
                      enum:
                        - INACTIVE
                    - type: string
                      enum:
                        - SUSPENDED
              required:
                - name
                - email
                - role
                - status
        required: true
      parameters:
        - schema:
            format: uuid
            type: string
          in: path
          name: id
          required: true
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    format: uuid
                    type: string
                  name:
                    type: string
                  email:
                    format: email
                    type: string
                  role:
                    anyOf:
                      - type: string
                        enum:
                          - ADMIN
                      - type: string
                        enum:
                          - MANAGER
                      - type: string
                        enum:
                          - OPERATOR
                      - type: string
                        enum:
                          - VIEWER
                  status:
                    anyOf:
                      - type: string
                        enum:
                          - ACTIVE
                      - type: string
                        enum:
                          - INACTIVE
                      - type: string
                        enum:
                          - SUSPENDED
                  createdAt:
                    format: date-time
                    type: string
                  updatedAt:
                    format: date-time
                    type: string
                required:
                  - id
                  - name
                  - email
                  - role
                  - status
                  - createdAt
                  - updatedAt
    patch:
      summary: Partially update a user
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  minLength: 2
                  maxLength: 120
                  type: string
                email:
                  format: email
                  type: string
                role:
                  anyOf:
                    - type: string
                      enum:
                        - ADMIN
                    - type: string
                      enum:
                        - MANAGER
                    - type: string
                      enum:
                        - OPERATOR
                    - type: string
                      enum:
                        - VIEWER
                status:
                  anyOf:
                    - type: string
                      enum:
                        - ACTIVE
                    - type: string
                      enum:
                        - INACTIVE
                    - type: string
                      enum:
                        - SUSPENDED
      parameters:
        - schema:
            format: uuid
            type: string
          in: path
          name: id
          required: true
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    format: uuid
                    type: string
                  name:
                    type: string
                  email:
                    format: email
                    type: string
                  role:
                    anyOf:
                      - type: string
                        enum:
                          - ADMIN
                      - type: string
                        enum:
                          - MANAGER
                      - type: string
                        enum:
                          - OPERATOR
                      - type: string
                        enum:
                          - VIEWER
                  status:
                    anyOf:
                      - type: string
                        enum:
                          - ACTIVE
                      - type: string
                        enum:
                          - INACTIVE
                      - type: string
                        enum:
                          - SUSPENDED
                  createdAt:
                    format: date-time
                    type: string
                  updatedAt:
                    format: date-time
                    type: string
                required:
                  - id
                  - name
                  - email
                  - role
                  - status
                  - createdAt
                  - updatedAt
    delete:
      summary: Delete a user
      tags:
        - Users
      parameters:
        - schema:
            format: uuid
            type: string
          in: path
          name: id
          required: true
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Default Response
...
servers:
  - url: /
    description: Current server
...